+++ packages/components/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @bubbles-ui/components
 
+## 1.1.69
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.1.69
+
+## 1.1.68
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.1.68
+
+## 1.1.67
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/icons@1.1.67
+
 ## 1.1.66
 
 ### Patch Changes