+++ packages/leemons/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @bubbles-ui/leemons
 
+## 1.1.69
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.1.69
+  - @bubbles-ui/calendars@1.1.69
+  - @bubbles-ui/editors@1.1.69
+  - @bubbles-ui/icons@1.1.69
+
+## 1.1.68
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.1.68
+  - @bubbles-ui/calendars@1.1.68
+  - @bubbles-ui/editors@1.1.68
+  - @bubbles-ui/icons@1.1.68
+
+## 1.1.67
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.1.67
+  - @bubbles-ui/calendars@1.1.67
+  - @bubbles-ui/editors@1.1.67
+  - @bubbles-ui/icons@1.1.67
+
 ## 1.1.66
 
 ### Patch Changes